+++ src/main/java/step3/Main.java
@@ -1,5 +1,8 @@
 package step3;
 
+import step3.domain.RacingGame;
+import step3.view.InputScanner;
+
 public class Main {
     public static void main(String[] args) {
         String inputNamesOfCars = InputScanner.getInputNamesOfCars();
