+++ src/main/java/step3/domain/RacingGame.java
@@ -1,4 +1,7 @@
-package step3;
+package step3.domain;
+
+import step3.domain.Car;
+import step3.lib.PrintMessage;
 
 import java.util.List;
 import java.util.stream.Collectors;
