+++ src/test/java/step3/RacingGameTest.java
@@ -3,6 +3,8 @@ package step3;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import step3.domain.Car;
+import step3.domain.RacingGame;
 
 import java.util.ArrayList;
 import java.util.List;
